refactor!: split EdgeStyle implementations and improve types
Introduce an `EdgeStyleFunction` type to precisely document the function accepted as perimeter. This
better guides implementers of `EdgeStyle` functions.

Start splitting `EdgeStyle` implementations. The "EdgeStyle" file was too large, the change also
prepare better tree-shaking in the future for people that don't want to use all default edgeStyle
implementations.
The `OrthConnector` and `ManhattanConnector` implementations are still located in the "EdgeStyle" file
because moving them require extra work. This could be done later.

Add various JSDoc improvements.

BREAKING CHANGES:
  - `CellStateStyle.loopStyle` no longer accept `Function`, but only the `EdgeStyleFunction`
  (limited impacts like on the `perimeter` property)
  - `Graph.createEdgeHandler` only accepts `EdgeStyleFunction` for the `edgeStyle` parameter.
  - Some internal methods of `EdgeStyle` are no longer available:
    - `scaleCellState`
    - `scalePointArray`
